% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_statis_phase1_chart.R
\name{plot_statis_phase1_chart_sum}
\alias{plot_statis_phase1_chart_sum}
\title{Plot Control Chart - Sum of Robust Mahalanobis Distances (Phase 1)}
\usage{
plot_statis_phase1_chart_sum(
  chi_sum_df,
  num_vars,
  title = "Robust STATIS Dual Control Chart - Phase 1 "
)
}
\arguments{
\item{chi_sum_df}{A data frame with columns \code{Batch} and \code{Robust_STATIS_Distance},
usually obtained by aggregating \code{phase1_result$standardized_data}.}

\item{num_vars}{Integer. Number of variables used in the multivariate analysis (to compute the Chi² threshold).}

\item{title}{Optional string. Plot title.}
}
\value{
A ggplot2 object.
}
\description{
Generates a control chart using the \strong{sum} of robust Mahalanobis distances for each batch,
based on the output from \code{robust_statis_phase1()}.
}
\examples{
data("datos_farma")
phase1_result <- robust_statis_phase1(
  subset(datos_farma, Status == "Under Control"),
  variables = c("Concentration", "Humidity", "Dissolution", "Density")
)
chi_sum_df <- aggregate(Robust_STATIS_Distance ~ Batch,
                        data = phase1_result$standardized_data, sum)
plot_statis_phase1_chart_sum(chi_sum_df, num_vars = 4)
}
