% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasimulation.R
\name{simulate_pharma_batches}
\alias{simulate_pharma_batches}
\title{Simulate Pharmaceutical Manufacturing Batches}
\usage{
simulate_pharma_batches(obs_por_lote = 10)
}
\arguments{
\item{obs_por_lote}{Integer. Number of observations (samples) per batch. Default is 10.}
}
\value{
A data frame with columns: Batch, Fase, Status, Concentration, Humidity, Dissolution, Density.
}
\description{
This function generates simulated data for a pharmaceutical manufacturing process.
It includes 10 in-control batches for Phase 1 and 7 additional batches for Phase 2,
where 4 are under control and 3 are out of control.
}
